refactor: correct spelling of 'deviceConstants' in imports across multiple files, adjust imports order in deviceConstants file
