+++ packages/vscode-extension/src/webview/components/InspectDataMenu.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import * as DropdownMenu from "@radix-ui/react-dropdown-menu";
 import { Frame, InspectDataStackItem } from "../../common/Project";
-import { DeviceProperties } from "../utilities/deviceContants";
+import { DeviceProperties } from "../utilities/deviceConstants";
 import "./InspectDataMenu.css";
 
 type OnSelectedCallback = (item: InspectDataStackItem) => void;
