+++ packages/vscode-extension/src/webview/views/ManageDevicesView.tsx
@@ -13,7 +13,7 @@ import Label from "../components/shared/Label";
 import Button from "../components/shared/Button";
 import { useProject } from "../providers/ProjectProvider";
 import { useModal } from "../providers/ModalProvider";
-import { mapIdToModel } from "../utilities/deviceContants";
+import { mapIdToModel } from "../utilities/deviceConstants";
 
 import "../components/shared/SwitchGroup.css";
 import { useStore } from "../providers/storeProvider";
