+++ packages/vscode-extension/src/webview/views/PreviewView.tsx
@@ -21,7 +21,7 @@ import {
   ZoomLevelType,
 } from "../../common/Project";
 import { Platform, useUtils } from "../providers/UtilsProvider";
-import { AndroidSupportedDevices, iOSSupportedDevices } from "../utilities/deviceContants";
+import { AndroidSupportedDevices, iOSSupportedDevices } from "../utilities/deviceConstants";
 import "./View.css";
 import "./PreviewView.css";
 import ReplayIcon from "../components/icons/ReplayIcon";
